Add basal smoothing support, remove rounding from initial recommendation.
When submitting a new job, the resulting basal suggestions
can be smoothed by supplying one of the supported smoothing
types ('none', 'low', 'medium', 'high'). The default is 'none'.

Also, the rounded recommendation is removed from the initial basal
recommendation and is now calculated on I/O edges such as database
and service endpoint.